# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Availability {
  year: number;
  month: number;
  occupiedDates: string[];
}

export interface ReservationRequestInput {
  /** ISO date string (YYYY-MM-DD) for the first night of the stay */
  checkIn: string;
  /** ISO date string (YYYY-MM-DD) for the departure day (exclusive end) */
  checkOut: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 8
     * @maxLength 30
     */
  phone: string;
}

export interface ReservationRequest {
  eventId: string;
  checkIn: string;
  checkOut: string;
  nights: number;
  name: string;
  phone: string;
  whatsappUrl: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailabilityParams = {
/**
 * @minimum 2020
 * @maximum 2100
 */
year: number;
/**
 * @minimum 1
 * @maximum 12
 */
month: number;
};