# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reservations
    description: Availability and reservation requests
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /availability:
    get:
      operationId: getAvailability
      tags: [reservations]
      summary: Get occupied dates for a month
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            minimum: 2020
            maximum: 2100
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        "200":
          description: Dates occupied by confirmed Calendar events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
        "502":
          description: Calendar provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservation-requests:
    post:
      operationId: createReservationRequest
      tags: [reservations]
      summary: Create a pre-reservation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationRequestInput"
      responses:
        "201":
          description: Pre-reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Date is no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Calendar provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Availability:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        occupiedDates:
          type: array
          items:
            type: string
      required:
        - year
        - month
        - occupiedDates
    ReservationRequestInput:
      type: object
      properties:
        checkIn:
          type: string
          description: ISO date string (YYYY-MM-DD) for the first night of the stay
        checkOut:
          type: string
          description: ISO date string (YYYY-MM-DD) for the departure day (exclusive end)
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 8
          maxLength: 30
      required:
        - checkIn
        - checkOut
        - name
        - phone
    ReservationRequest:
      type: object
      properties:
        eventId:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        nights:
          type: integer
        name:
          type: string
        phone:
          type: string
        whatsappUrl:
          type: string
      required:
        - eventId
        - checkIn
        - checkOut
        - nights
        - name
        - phone
        - whatsappUrl
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error