# lib/api-zod/src/generated/types/reservationRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationRequest {
  eventId: string;
  checkIn: string;
  checkOut: string;
  nights: number;
  name: string;
  phone: string;
  whatsappUrl: string;
}