# lib/api-zod/src/generated/types/reservationRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationRequestInput {
  /** ISO date string (YYYY-MM-DD) for the first night of the stay */
  checkIn: string;
  /** ISO date string (YYYY-MM-DD) for the departure day (exclusive end) */
  checkOut: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 8
     * @maxLength 30
     */
  phone: string;
}